♻️ HSNCNJ init: decompose execution planning
