+++ .agentplane/tasks/202605290115-HSNCNJ/pr/github-body.md
@@ -1,0 +1,33 @@
+Task: `202605290115-HSNCNJ`
+Title: Init execution decomposition
+Canonical task record: `.agentplane/tasks/202605290115-HSNCNJ/README.md`
+
+## Summary
+
+Init execution decomposition
+
+Decompose packages/agentplane/src/cli/run-cli/commands/init/execution.ts by extracting focused init planning/effect helpers while preserving init behavior and reducing the runtime hotspot warning count.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/cli/run-cli/commands/init/execution.ts by extracting focused init planning/effect helpers while preserving init behavior and reducing the runtime hotspot warning count.
+- Out of scope: unrelated refactors not required for "Init execution decomposition".
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T01:16:22.379Z
+- Branch: task/202605290115-HSNCNJ/init-execution-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
