+++ .agentplane/tasks/202605290115-HSNCNJ/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 HSNCNJ task: Init execution decomposition [202605290115-HSNCNJ]
